Update: 共通コンポーネントを別コンポーネントに切り出す
