+++ src/App.tsx
@@ -1,34 +1,15 @@
 import "./lib/tailwind.css";
 
 import { MantineProvider } from "@mantine/core";
-import FooterLinks from "components/FooterLinks";
 import GameManagement from "components/GameManagement/GameManagement";
-import HeaderAction from "components/HeaderAction";
-
-const data = [
-  {
-    link: "#",
-    label: "プライバシーポリシー",
-  },
-  {
-    link: "#",
-    label: "利用規約",
-  },
-  {
-    link: "#",
-    label: "お問い合わせ",
-  },
-];
-
-const isLogin = true;
-const user = { name: "User", image: "user-image-url" };
+import MainLayout from "components/Layout/MainLayout";
 
 const App = () => {
   return (
     <MantineProvider withGlobalStyles withNormalizeCSS>
-      <HeaderAction isLogin={isLogin} user={user} />
-      <GameManagement />
-      <FooterLinks {...{ data }} />
+      <MainLayout>
+        <GameManagement />
+      </MainLayout>
     </MantineProvider>
   );
 };
